refactoring tests - move InvokeScriptTransactionTestData to transactions package
