+++ src/test/java/com/wavesplatform/wavesj/json/deser/InvokeScriptTransactionDeserTest.java
@@ -1,6 +1,6 @@
 package com.wavesplatform.wavesj.json.deser;
 
-import com.wavesplatform.wavesj.json.InvokeScriptTransactionTxTestData;
+import com.wavesplatform.wavesj.transactions.InvokeScriptTransactionTestData;
 import com.wavesplatform.wavesj.transactions.InvokeScriptTransaction;
 import org.junit.Test;
 
@@ -11,24 +11,24 @@ public class InvokeScriptTransactionDeserTest extends TransactionDeserTest {
     @Test
     public void V1DeserializeTest() throws IOException {
         deserializationTest(
-                InvokeScriptTransactionTxTestData.txFullJson(),
-                InvokeScriptTransactionTxTestData.txFull(),
+                InvokeScriptTransactionTestData.txFullJson(),
+                InvokeScriptTransactionTestData.txFull(),
                 InvokeScriptTransaction.class);
     }
 
     @Test
     public void testV1DeserTxNoFunctionCall() throws IOException {
         deserializationTest(
-                InvokeScriptTransactionTxTestData.txNoFunctionCallJson(),
-                InvokeScriptTransactionTxTestData.txNoFunctionCall(),
+                InvokeScriptTransactionTestData.txNoFunctionCallJson(),
+                InvokeScriptTransactionTestData.txNoFunctionCall(),
                 InvokeScriptTransaction.class);
     }
 
     @Test
     public void testV1DeserTxNoPayment() throws IOException {
         deserializationTest(
-                InvokeScriptTransactionTxTestData.txNoPaymentJson(),
-                InvokeScriptTransactionTxTestData.txNoPayment(),
+                InvokeScriptTransactionTestData.txNoPaymentJson(),
+                InvokeScriptTransactionTestData.txNoPayment(),
                 InvokeScriptTransaction.class);
     }
